更新工程文件，提交，同时生成的exe文件不再上传到库中

+++ klib/kgl_smooth_normals_generator.cc
@@ -48,8 +48,8 @@ namespace kgl
         {
             x = x < 0 ? 0 : x;
             z = z < 0 ? 0 : z;
-            x = x >= heights.size2() ? heights.size2() - 1 : x;
-            z = z >= heights.size1() ? heights.size1() - 1 : z;
+			x = x >= static_cast<int>(heights.size2()) ? static_cast<int>(heights.size2() - 1) : x;
+			z = z >= static_cast<int>(heights.size1()) ? static_cast<int>(heights.size1() - 1) : z;
             return heights(z, x);// [z][x];
         }
     }
